refactor: M2 — move LLM adapter contract to shared, delete SessionTransport ghost
T15: moved LLMAdapter/LLMResponse + the full LLM call contract to @covel/shared; runtime
  re-exports for backward compat; create no longer depends on @covel/runtime — the only
  wrong-layer edge in the dependency graph is gone.
T19: deleted the SessionTransport interface + the never-wired client->server command cluster
  (ClientInfo/ServerCapabilities/SessionCommand/...); rewrote protocol.md to honestly describe
  the real transport (fetch REST + two SSE streams + DataService data layer) instead of an
  unimplemented transport-agnostic abstraction.
