+++ packages/shared/src/types/index.ts
@@ -18,6 +18,20 @@ export type {
 } from "./llm-content-parts.js";
 
 export type {
+  LLMTextPart,
+  LLMImagePart,
+  LLMContentPart,
+  LLMMessageContent,
+  LLMMessage,
+  LLMToolCall,
+  LLMResponse,
+  LLMToolDefinition,
+  LLMResponseFormat,
+  LLMStreamEvent,
+  LLMAdapter,
+} from "./llm-adapter.js";
+
+export type {
   WorldDataSourceKind,
   WorldDataMergeMode,
   WorldDataEffect,
@@ -173,14 +187,6 @@ export type {
 export { PROPOSAL_TYPES } from "./proposal.js";
 
 export type {
-  ClientCapabilities,
-  ServerCapabilities,
-  CommandType,
-  SessionCommand,
-  SessionCreatePayload,
-  SessionRestorePayload,
-  TurnSubmitPayload,
-  InputSubmitPayload,
   ProtocolEventType,
   ProtocolEvent,
   SessionSnapshot,
@@ -188,8 +194,6 @@ export type {
   SnapshotCharacter,
   SnapshotTraceEvent,
   SnapshotPluginStatus,
-  SessionTransport,
-  ClientInfo,
 } from "./protocol.js";
 
 export type {
